Write tests descriptions

+++ tests/test_A0_initialization.py
@@ -8,79 +8,92 @@ import argparse
 import src.A0_initialization as initializing
 from mvs_eland_tool.mvs_eland_tool import main
 
-from .constants import REPO_PATH
-
-OUTPUT_PATH = os.path.join(".", "tests", "other")
+from .constants import REPO_PATH, CSV_ELEMENTS, INPUTS_COPY
 
 
-class TestOutputPath:
-
-    output_path = os.path.join(".", "tests", "MVS_outputs")
-
-    def setup_method(self):
-
-        os.mkdir(self.output_path)
-        os.mkdir(os.path.join(self.output_path, "dummy"))
-
-    def test_create_output_path_ask_to_overwrite_existing_no(self):
-        with mock.patch("builtins.input", return_value="no"):
-            with pytest.raises(FileExistsError):
-                initializing.check_output_directory(
-                    os.path.join(self.output_path), overwrite=False
-                )
-            assert os.path.exists(self.output_path)
-            assert os.path.exists(os.path.join(self.output_path, "dummy"))
-
-    def test_create_output_path_ask_overwrite_existing_yes(self):
-        with mock.patch("builtins.input", return_value="yes"):
-            initializing.check_output_directory(
-                os.path.join(self.output_path), overwrite=False
-            )
-            assert os.path.exists(self.output_path)
-            assert not os.path.exists(os.path.join(self.output_path, "dummy"))
-
-    def test_create_output_path_existing_yes(self):
-        initializing.check_output_directory(
-            os.path.join(self.output_path), overwrite=True
-        )
-        assert os.path.exists(self.output_path)
-        assert not os.path.exists(os.path.join(self.output_path, "dummy"))
-
-    def teardown_method(self):
-        shutil.rmtree(self.output_path, ignore_errors=True)
-
-
-class TestUserInput:
+class TestProcessUserArguments:
 
     output_path = os.path.join(".", "tests", "MVS_outputs")
     input_path = os.path.join("tests", "inputs")
     fake_input_path = os.path.join("tests", "fake_inputs")
 
-    def test_user_input_path_folder_copy_in_output(self):
-        initializing.get_user_input(path_output_folder=self.output_path)
-        assert os.path.exists(self.output_path)
+    def test_input_folder_is_copied_in_output_within_folder_named_input(self):
+        pass
+        # initializing.get_user_input(path_output_folder=self.output_path)
+        # assert os.path.exists(self.output_path)
 
-    def test_user_input_path_input_folder_not_existing(self):
-        with pytest.raises(NotADirectoryError):
-            initializing.get_user_input(
-                path_input_file=self.fake_input_path,
-                path_output_folder=self.output_path,
-            )
+    def test_input_folder_not_existing_raise_filenotfound_error(self):
+        pass
+        # with pytest.raises(NotADirectoryError):
+        #     initializing.get_user_input(
+        #         path_input_file=self.fake_input_path,
+        #         path_output_folder=self.output_path,
+        #     )
 
-    def test_user_input_path_input_file_not_existing(self):
-        with pytest.raises(FileNotFoundError):
-            initializing.get_user_input(
-                path_input_file=os.path.join(self.input_path, "not_existing.json"),
-                path_output_folder=self.output_path,
-                overwrite=True,
-            )
+    def test_if_json_opt_and_no_json_file_in_input_folder_raise_filenotfound_error(
+        self,
+    ):
+        pass
+        # with pytest.raises(FileNotFoundError):
+        #     initializing.get_user_input(
+        #         path_input_file=os.path.join(self.input_path, "not_existing.json"),
+        #         path_output_folder=self.output_path,
+        #         overwrite=True,
+        #     )
+
+    def test_if_json_opt_and_more_than_one_json_file_in_input_folder_raise_fileexists_error(
+        self,
+    ):
+        pass
+        # with pytest.raises(FileNotFoundError):
+        #     initializing.get_user_input(
+        #         path_input_file=os.path.join(self.input_path, "not_existing.json"),
+        #         path_output_folder=self.output_path,
+        #         overwrite=True,
+        #     )
+
+    def test_if_json_opt_path_input_file_set_to_path_input_folder(self):
+        """Check that the path_input_file is <path_input_folder>
+
+            in module A1_csv_to_json this will be transformed into a .json file
+        """
+        pass
+
+    def test_if_csv_opt_and_csv_elements_folder_not_in_input_folder_raise_filenotfound_error(
+        self,
+    ):
+        pass
+        # with pytest.raises(FileNotFoundError):
+        #     initializing.get_user_input(
+        #         path_input_file=os.path.join(self.input_path, "not_existing.json"),
+        #         path_output_folder=self.output_path,
+        #         overwrite=True,
+        #     )
+
+    def test_if_csv_opt_path_input_file_set_to_path_input_folder_mvs_csv_config_dot_json(
+        self,
+    ):
+        """Check that the path_input_file is <path_input_folder>/mvs_csv_config.json """
+        pass
+
+    def test_if_log_opt_display_output_is_set_with_correct_value(self):
+        pass
+
+    def test_if_path_output_folder_exists_raise_fileexists_error(self):
+        """The error message should advice the user to use -f option to force overwrite"""
+        pass
+
+    def test_if_f_opt_preexisting_path_output_folder_should_be_replaced(self):
+        pass
 
     def teardown_method(self):
-        if os.path.exists(self.output_path):
-            shutil.rmtree(self.output_path, ignore_errors=True)
+        pass
+        # if os.path.exists(self.output_path):
+        #     shutil.rmtree(self.output_path, ignore_errors=True)
 
 
 def test_check_input_path_posix():
+    """Verify the code works on both windows and linux path systems"""
     if os.name == "posix":
         folder = initializing.check_input_directory(
             "{}/inputs/working_example.json".format(REPO_PATH)
@@ -93,13 +106,20 @@ def test_check_input_path_posix():
     assert folder == os.path.join(REPO_PATH, "inputs")
 
 
+OUTPUT_PATH = os.path.join(".", "tests", "other")
+
+
 class TestCommandLineInput:
 
     parser = initializing.create_parser()
 
-    def test_input_file(self):
-        parsed = self.parser.parse_args(["-i", "input_file"])
+    def test_input_json_file(self):
+        parsed = self.parser.parse_args(["-json", "input_file"])
         assert parsed.path_input_file == "input_file"
+
+    def test_input_csv_file(self):
+        parsed = self.parser.parse_args(["-csv", "input_file"])
+        assert parsed.path_input_file == os.path.join("input_file", CSV_ELEMENTS)
 
     def test_output_folder(self):
         parsed = self.parser.parse_args(["-o", "output_folder"])
@@ -140,4 +160,5 @@ class TestCommandLineInput:
         assert os.path.exists(OUTPUT_PATH)
 
     def teardown_method(self):
-        shutil.rmtree(OUTPUT_PATH, ignore_errors=True)
+        if os.path.exists(OUTPUT_PATH):
+            shutil.rmtree(OUTPUT_PATH, ignore_errors=True)
